Move exempt from 2sv tests into shared manage 2sv test
